ID-B-5.2-work-with-comments
