+++ backend/FindYourFriendAmongPets/FindYourFriendAmongPets.Application/Volunteers/Create/FullNameDto.cs
@@ -1,0 +1,6 @@
+﻿namespace FindYourFriendAmongPets.Application.Volunteers.Create;
+
+public record FullNameDto(
+    string FirstName,
+    string LastName,
+    string? Patronymic);
